Add files via upload
This is the first version which includes network communication between Coopener (RasPi) and SmartHome (Linux web server). The protocol is likely extremely buggy and is definitely not secure (don't use over a public network). The web interface is minimalistic but includes two way functionality.
The Arduino code hasn't changed since the last version (and is likely not going to change for a while), but it is included in this update again for completeness